# vyora-api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Vyora API
  version: 1.0.0-draft.1
  summary: Authenticated party and event-sync contract for the Vyora credit ledger.
  description: |
    Design-only contract. **No implementation of this API exists.**

    Three rules govern every path below.

    **1. The tenant is resolved from the token, never from the request.**
    A `Merchant` is a server-owned workspace identified by a UUID. No path,
    query parameter, header or request body in this contract accepts a
    `merchantId`. A caller cannot name, select or override the workspace they
    are reading; the access token decides, and the server scopes every query.

    **2. `Party` is the only contact resource.**
    There is no customer or supplier entity, and no party carries a role field.
    Direction lives on the financial entry, not the contact — see
    `docs/platform/adr/0002-party-is-the-core-domain-entity.md`.

    **3. Sync transfers events, never state.**
    `POST /api/v1/sync/push` appends immutable client events; the server never
    overwrites. `GET /api/v1/sync/pull` is cursor-based over server receipt
    order. There is no endpoint that replaces a merchant's ledger with a
    submitted state document.
  contact:
    name: Vyora Platform
  license:
    name: Proprietary — EasyTechLabs
    identifier: LicenseRef-EasyTechLabs-Proprietary
servers:
  - url: http://127.0.0.1:4000
    description: |
      Local development only. **No deployed environment exists**, and no
...
security:
  - bearerAuth: []
  - devAuth: []
paths:
...
  /api/v1/parties/{partyId}/credits:
    parameters:
      - $ref: "#/components/parameters/PartyId"
    post:
      operationId: recordCredit
      tags: [Ledger]
      summary: Record a credit entry against a party.
      description: |
        Appends one immutable `CreditRecorded` event. The request carries the
        entry, not a new state for the party — there is no field here that
        replaces anything, and the party's balance is never sent or stored.

        `kind` is the direction, from the merchant's point of view. `given`
        means the party owes the merchant more; `taken` means the merchant owes
        the party more. The same party can be on either side on different days,
        which is why no role is stored on the contact.

        The entry id is client-minted so the entry exists offline the moment the
        merchant records it, and it doubles as the natural deduplication key
        alongside `Idempotency-Key`.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordCreditRequest"
      responses:
        "201":
          description: |
            Credit recorded; the event is in the log.

            **An idempotent replay returns this same `201` and the same body,
            and appends no second event.** The stored response is returned
            verbatim — status included — because a retry that is
            distinguishable from the original is not idempotent. A client whose
            connection dropped mid-write cannot tell whether it or its retry
            recorded the entry, and it does not need to: exactly one entry and
            exactly one event exist either way.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            The `Idempotency-Key` was reused with a different body, or the entry
            id already exists with different content. Never an overwrite.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/parties/{partyId}/statement:
    parameters:
      - $ref: "#/components/parameters/PartyId"
    get:
      operationId: getPartyStatement
      tags: [Ledger]
      summary: A party's statement and derived balance.
      description: |
        The dispute-settling history for one party: every entry, oldest first,
        each with the running outstanding after it.

        **Entirely derived.** Rows are folded from accepted events at read time
        and nothing here is stored as a total. If the projection and the log
        ever disagree, the log wins and the projection is rebuilt.

        `404` when the party does not exist **or belongs to another workspace** —
        deliberately indistinguishable, so a caller cannot probe for another
        merchant's parties.
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: The statement.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartyStatement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/parties/{partyId}/payments:
    parameters:
      - $ref: "#/components/parameters/PartyId"
    post:
      operationId: recordPayment
      tags: [Ledger]
      summary: Record a payment against a party.
      description: |
        Appends one immutable `PaymentRecorded` event — the same type and
        payload already in log format v2 on merchants' devices.

        **A payment does not settle a specific credit.** It moves the party's
        net position and nothing else. Real udhaar is not invoice-matched: a
        merchant takes ₹500 against a running tab, not against the entry from
        the 14th. Allocating payments to entries would invent a relationship
        the merchant never asserted, and would then have to be un-invented when
        an older entry arrived late from another device.

        `kind` is the direction from the merchant's point of view. `received` —
        the party paid the merchant, so they owe less. `paid` — the merchant
        paid the party, so the merchant owes less.

        Recording a payment cannot take a balance past zero into a fabricated
        one: over-payment is permitted and simply inverts the position, because
        refusing it would make the ledger disagree with what physically
        happened at the counter.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordPaymentRequest"
      responses:
        "201":
          description: |
            Payment recorded; the event is in the log.

            **An idempotent replay returns this same `201` and the same body,
            and appends no second event.** The stored response is returned
            verbatim — status included — because a retry that is
            distinguishable from the original is not idempotent.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            The `Idempotency-Key` was reused with a different body, or the entry
            id already exists with different content. Never an overwrite.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/parties/{partyId}/summary:
    parameters:
      - $ref: "#/components/parameters/PartyId"
    get:
      operationId: getPartyLedgerSummary
      tags: [Ledger]
      summary: A party's ledger totals, without the entries.
      description: |
        What the statement adds up to, for callers that need the totals but not
        the history — a party card, a list row, a mobile screen on a slow
        connection.

        **Every number here is folded from accepted events at read time.**
        Nothing is stored and nothing is cached. This endpoint and
        `/statement` derive their balance through the same fold, so a summary
        can never disagree with the statement it summarises.

        The four `totals` are unsigned gross amounts per direction; `balance.net`
        is the single signed number. They are reported separately because
        "you've given ₹40,000 and been paid ₹38,500" is a different fact from
        "₹1,500 outstanding", and a merchant chasing a debt needs both.

        `404` when the party does not exist **or belongs to another workspace** —
        deliberately indistinguishable.
      responses:
        "200":
          description: The summary.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartyLedgerSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Production scheme. The token carries the merchant workspace, the user
        and the device. The server resolves tenant scope from it exclusively.

        Clients never send `merchantId`. There is no request field anywhere in
        this contract through which a caller could name a workspace, so a
        token-scoping bug cannot be escalated by a crafted request.

    devAuth:
      type: apiKey
      in: header
      name: X-Vyora-Dev-Identity
      description: |
        **Development and contract testing only.**

        Accepted only when the server is started with `VYORA_DEV_AUTH=true`,
        which is refused outright when `NODE_ENV=production`. The value names a
        seeded synthetic workspace from the local fixture set; it can only ever
        resolve to a fixture merchant, never to a real one.

        This is a *distinct scheme*, not a bearer token with a magic value, so
        that its acceptance is a single explicit branch — greppable, testable,
        and impossible to reach by accident in a production build. See
        `docs/platform/auth-and-tenant-model.md` §4.
  headers:
    XRequestId:
      description: Correlation id echoed on every response, success or failure.
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
  parameters:
    PartyId:
      name: partyId
      in: path
      required: true
      description: Client-minted party identifier.
      schema:
        $ref: "#/components/schemas/PartyRef"
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum items per page.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUID scoping this request for replay. Retrying with
        the same key and an identical body returns the original result; reusing
        it with a different body is a `409`.
      schema:
        $ref: "#/components/schemas/Uuid"
...
  responses:
    BadRequest:
      description: Malformed request — syntax, parameters or headers.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing, malformed or expired credentials.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: |
        No such resource in the caller's workspace. Returned identically
        whether the resource is absent or belongs to another merchant.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationFailed:
      description: Well-formed request whose contents failed field-level validation.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    Uuid:
      type: string
      format: uuid
      description: Server-minted identifier. Always a bare RFC 4122 UUID.
    Timestamp:
      type: string
      format: date-time
      description: ISO-8601 instant with an explicit offset, e.g. `2026-08-05T09:14:22.115Z`.
    BusinessDate:
      type: string
      format: date
      description: |
        The merchant's trading date, `YYYY-MM-DD`, in their local timezone.
        Distinct from a timestamp: a sale on the evening of the 4th belongs to
        the 4th's book regardless of UTC offset.
    Money:
      type: integer
      minimum: 0
      description: |
        Rupees as a non-negative integer. Direction is carried by the entry's
        `kind`, never by the sign of the amount, so an amount is never
        negative anywhere in this contract.
    VyoraId:
      type: string
      pattern: "^(pty|txn|pay|evt)_[A-Za-z0-9-]{8,64}$"
      description: |
        Client-minted domain identifier: a type prefix, an underscore, then a
        UUID (`pty_1f0c…`). The prefix makes an id self-describing in a log
        line and lets entry ids stay unique across credits and payments.

        The UUID fallback for pre-`crypto.randomUUID` runtimes emits a shorter
        base-36 suffix, which is why the pattern is a length range rather than
        a strict UUID. **This format already exists on merchants' devices and
        cannot be changed without rewriting their event logs**, so the contract
        accepts it as-is rather than forcing a migration.
    PartyRef:
      allOf:
        - $ref: "#/components/schemas/VyoraId"
      pattern: "^pty_[A-Za-z0-9-]{8,64}$"
      description: A `VyoraId` restricted to the party prefix.
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorObject"
    ErrorObject:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: |
            Developer-facing English. **Never rendered to a merchant.** The
            client owns merchant-facing wording, which is already written,
            translated and tone-checked on the device.
        requestId:
          allOf:
            - $ref: "#/components/schemas/Uuid"
          description: Matches the `X-Request-Id` response header. Quote it in bug reports.
        details:
          type: array
          description: Present when the failure is attributable to specific fields.
          items:
            $ref: "#/components/schemas/FieldError"
...
    ErrorCode:
      type: string
      description: Machine-readable failure class. Stable; new values may be added.
      enum:
        - BAD_REQUEST
        - UNAUTHENTICATED
        - TOKEN_EXPIRED
        - FORBIDDEN
        - NOT_FOUND
        - VALIDATION_FAILED
        - VERSION_CONFLICT
        - PRECONDITION_REQUIRED
        - IDEMPOTENCY_KEY_REUSED
        - RESOURCE_ALREADY_EXISTS
        - PAYLOAD_TOO_LARGE
        - RATE_LIMITED
        - SYNC_CURSOR_INVALID
        - SYNC_CURSOR_EXPIRED
        - SYNC_CLIENT_TOO_OLD
        - SYNC_SCHEMA_VERSION_UNSUPPORTED
        - DEPENDENCY_UNAVAILABLE
        - INTERNAL_ERROR

    # ── System ──────────────────────────────────────────────────────────────
...
    PartyPosition:
      type: string
      enum: [owes_merchant, merchant_owes, settled, no_entries]
      description: |
        **Derived, never stored.** Computed by folding entry direction at read
        time. A party is not "a customer" or "a supplier"; they hold a position
        that can invert with the next entry.
...
    PartyBalance:
      type: object
      additionalProperties: false
      required: [net, position, entryCount]
      readOnly: true
      description: |
        Derived projection of this party's entries. **No balance is stored
        anywhere** — server or device. A stored balance would be a second
        source of truth and would drift the instant a late offline event
        arrived, which offline-first makes routine rather than exceptional.
      properties:
        net:
          type: integer
          description: |
            Signed rupees. Positive means the party owes the merchant;
            negative means the merchant owes the party. This is the only
            signed amount in the contract, and it is computed, never submitted.
        position:
          $ref: "#/components/schemas/PartyPosition"
        entryCount:
          type: integer
          minimum: 0
        lastActivityAt:
          type: [string, "null"]
          format: date-time
...
    RecordCreditRequest:
      type: object
      additionalProperties: false
      required: [id, amount, kind, date]
      description: |
        Note what is absent: no `merchantId`, no `partyId` in the body (it is in
        the path), and no balance. The workspace comes from the token and the
        balance is derived.
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/VyoraId"
          description: Client-minted entry id, so the entry exists offline immediately.
        amount:
          $ref: "#/components/schemas/Money"
        kind:
          $ref: "#/components/schemas/EntryKind"
        description:
          type: [string, "null"]
          maxLength: 500
        date:
          $ref: "#/components/schemas/BusinessDate"
        dueDate:
          type: [string, "null"]
          format: date
        createdAt:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
          description: |
            When the merchant actually recorded it. Preserved so an entry made
            offline keeps its true moment rather than the moment it reached a
            server.
    RecordPaymentRequest:
      type: object
      additionalProperties: false
      required: [id, amount, kind, date]
      description: |
        Mirrors `RecordCreditRequest` deliberately: same absences (no
        `merchantId`, no `partyId` in the body, no balance), same client-minted
        id, same offline-first reasoning.

        There is no `dueDate` and no `appliesTo`. A payment has nothing falling
        due, and it settles the running position rather than a nominated entry.
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/VyoraId"
          description: Client-minted entry id, so the payment exists offline immediately.
        amount:
          $ref: "#/components/schemas/Money"
        kind:
          $ref: "#/components/schemas/PaymentKind"
        note:
          type: [string, "null"]
          maxLength: 500
          description: |
            The merchant's own words — "part payment", "cash at shop". Named
            `note` rather than `description` because that is what the field is
            called on the local `Payment` and renaming it in transit would make
            the two records look like different things.
        date:
          $ref: "#/components/schemas/BusinessDate"
        createdAt:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
          description: |
            When the merchant actually recorded it. Preserved so a payment
            taken offline keeps its true moment.
    LedgerEntry:
      type: object
      additionalProperties: false
      required: [id, partyId, entryType, direction, amount, date, createdAt, eventId]
      description: One accepted entry, as stored. Read-only.
      properties:
        id:
          $ref: "#/components/schemas/VyoraId"
        partyId:
          $ref: "#/components/schemas/PartyRef"
        entryType:
          type: string
          enum: [credit, payment]
          description: |
            `credit` from `POST …/credits`, `payment` from `POST …/payments`.
            Both are entries in one log; the distinction is what the merchant
            did, not where it is stored.
        direction:
          type: string
          enum: [given, taken, received, paid]
        amount:
          $ref: "#/components/schemas/Money"
        description:
          type: [string, "null"]
        date:
          $ref: "#/components/schemas/BusinessDate"
        dueDate:
          type: [string, "null"]
          format: date
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        eventId:
          allOf:
            - $ref: "#/components/schemas/VyoraId"
          description: The immutable event this entry came from. The audit handle.
    StatementRow:
      allOf:
        - $ref: "#/components/schemas/LedgerEntry"
        - type: object
          required: [signedAmount, runningNet, label]
          properties:
            signedAmount:
              type: integer
              description: |
                This entry's effect on "they owe me": `+given`, `+paid`,
                `−taken`, `−received`.
            runningNet:
              type: integer
              description: The outstanding after this entry, folded oldest-first.
            label:
              type: string
              description: Developer-facing, e.g. `Credit given`. The client owns merchant wording.
    PartyStatement:
      type: object
      additionalProperties: false
      required: [partyId, rows, balance]
      properties:
        partyId:
          $ref: "#/components/schemas/PartyRef"
        rows:
          type: array
          description: Oldest first, so `runningNet` reads as a ledger does.
          items:
            $ref: "#/components/schemas/StatementRow"
        balance:
          $ref: "#/components/schemas/PartyBalance"
    LedgerTotals:
      type: object
      additionalProperties: false
      required: [creditGiven, creditTaken, paymentReceived, paymentPaid]
      readOnly: true
      description: |
        Gross unsigned amounts per direction. Deliberately not netted against
        each other here — `PartyBalance.net` is the one signed number, and
        keeping the gross figures separate is what lets a merchant see that a
        quiet ₹1,500 balance sits on ₹40,000 of trade rather than ₹1,500 of it.
      properties:
        creditGiven:
          allOf:
            - $ref: "#/components/schemas/Money"
          description: Total credit the merchant handed over.
        creditTaken:
          allOf:
            - $ref: "#/components/schemas/Money"
          description: Total credit the merchant received.
        paymentReceived:
          allOf:
            - $ref: "#/components/schemas/Money"
          description: Total the party paid the merchant.
        paymentPaid:
          allOf:
            - $ref: "#/components/schemas/Money"
          description: Total the merchant paid the party.
    LedgerCounts:
      type: object
      additionalProperties: false
      required: [credits, payments]
      readOnly: true
      properties:
        credits:
          type: integer
          minimum: 0
        payments:
          type: integer
          minimum: 0
    PartyLedgerSummary:
      type: object
      additionalProperties: false
      required: [partyId, balance, totals, counts]
      readOnly: true
      description: |
        The statement's arithmetic without its rows. Folded from the same
        entries through the same code path as `/statement`, so the two cannot
        report different balances for the same party.
      properties:
        partyId:
          $ref: "#/components/schemas/PartyRef"
        balance:
          $ref: "#/components/schemas/PartyBalance"
        totals:
          $ref: "#/components/schemas/LedgerTotals"
        counts:
          $ref: "#/components/schemas/LedgerCounts"
        firstActivityAt:
          type: [string, "null"]
          format: date-time
          description: |
            When this relationship started. `null` when the party has no
            entries — a contact saved ahead of the first sale, which is normal.

    # ── Events ──────────────────────────────────────────────────────────────
...
    EntryKind:
      type: string
      enum: [given, taken]
      description: |
        Credit direction from the merchant's point of view. `given` — the
        merchant handed over goods or money, so the party owes them. `taken` —
        the merchant received on credit, so they owe the party.

        **This is where role lives.** The same party can be on either side on
        different days, which is exactly why no role is stored on the contact.
    PaymentKind:
      type: string
      enum: [received, paid]
      description: |
        Payment direction from the merchant's point of view. `received` — the
        party paid the merchant. `paid` — the merchant paid the party.
...